// src/models/enums/AuthSessionState.ts
export enum AuthSessionState {
    CIC_SESSION_CREATED = 'CIC_SESSION_CREATED',
    CIC_DATA_RECEIVED = 'CIC_DATA_RECEIVED',
    CIC_AUTH_CODE_ISSUED = 'CIC_AUTH_CODE_ISSUED',
    CIC_ACCESS_TOKEN_ISSUED = 'CIC_ACCESS_TOKEN_ISSUED',
    CIC_SESSION_ABORTED = 'CIC_SESSION_ABORTED',
    CIC_SESSION_FINISHED = 'CIC_SESSION_FINISHED'
}